+++ components/SSI/DIDxWallet/DIDOperations/NewDoc/index.tsx
@@ -1,12 +1,10 @@
 import { useStore } from "effector-react";
 import React, { useState, useCallback } from "react";
 import { toast } from "react-toastify";
-import { useRouter } from "next/router";
 import { $donation } from "../../../../../src/store/donation";
 import { $user } from "../../../../../src/store/user";
-import { updateIsController } from "../../../../../src/store/controller";
 import * as tyron from "tyron";
-import { SubmitCreate, Donate, SubmitRecover, Headline } from "../../../..";
+import { SubmitCreate, Donate, SubmitRecover } from "../../../..";
 import styles from "./styles.module.scss";
 
 interface InputType {
@@ -21,7 +19,6 @@ function Component(props: InputType) {
     }
   }, []);
 
-  const Router = useRouter();
   const user = useStore($user);
   const donation = useStore($donation);
 
@@ -327,24 +324,6 @@ function Component(props: InputType) {
   const did_services = services__.concat(services2);
   return (
     <>
-      <div className={styles.headlineWrapper}>
-        <Headline />
-      </div>
-      <div style={{ textAlign: 'left' }}>
-        <button
-          type="button"
-          className="button"
-          onClick={() => {
-            updateIsController(true);
-            Router.push(`/${user?.name}/xwallet/did`)
-          }}
-        >
-          <p style={{ color: 'silver' }}>operations menu</p>
-        </button>
-      </div>
-      <h2 style={{ color: '#ffff32', margin: '7%' }}>
-        DID {typeInput}
-      </h2>
       <div>
         {hideDoc &&
           <div style={{ margin: '7%', textAlign: 'center' }}>
