+++ pages/[username]/xwallet/did/create.tsx
@@ -15,25 +15,25 @@ function Create() {
       <Layout>
         <div className={styles.headlineWrapper}>
           <Headline />
-          <div style={{ textAlign: 'left' }}>
-            <button
-              type="button"
-              className="button"
-              onClick={() => {
-                updateIsController(true);
-                Router.push(`/${username}/xwallet/did`)
-              }}
-            >
-              <p style={{ color: 'silver' }}>operations menu</p>
-            </button>
-          </div>
-          <h2 style={{ color: '#ffff32', margin: '7%' }}>
-            DID create
-          </h2>
-          <h4>
-            With this transaction, you will generate a globally unique Decentralized Identifier (DID) and its DID Document.
-          </h4>
         </div>
+        <div style={{ textAlign: 'left' }}>
+          <button
+            type="button"
+            className="button"
+            onClick={() => {
+              updateIsController(true);
+              Router.push(`/${username}/xwallet/did`)
+            }}
+          >
+            <p style={{ color: 'silver' }}>operations menu</p>
+          </button>
+        </div>
+        <h2 style={{ color: '#ffff32', margin: '7%' }}>
+          DID create
+        </h2>
+        <h4 style={{width: '50%'}}>
+          With this transaction, you will generate a globally unique Decentralized Identifier (DID) and its DID Document.
+        </h4>
         <NewDoc typeInput="create" />
       </Layout>
     </>
