+++ pages/[username]/xwallet/did/social.tsx
@@ -1,10 +1,35 @@
+import { useRouter } from "next/router";
 import Layout from "../../../../components/Layout";
-import { DIDxWallet, DIDOperations, DidSocialRecovery } from "../../../../components";
+import { DidSocialRecovery, Headline } from "../../../../components";
+import { updateIsController } from "../../../../src/store/controller";
+import { useStore } from "effector-react";
+import { $user } from "../../../../src/store/user";
+import styles from "./styles.module.scss";
 
 function Social() {
+  const Router = useRouter();
+  const username = useStore($user)?.name;
   return (
     <>
       <Layout>
+        <div className={styles.headlineWrapper}>
+          <Headline />
+        </div>
+        <div style={{ textAlign: 'left' }}>
+          <button
+            type="button"
+            className="button"
+            onClick={() => {
+              updateIsController(true);
+              Router.push(`/${username}/xwallet/did`)
+            }}
+          >
+            <p style={{ color: 'silver' }}>operations menu</p>
+          </button>
+        </div>
+        <h2 style={{ color: '#ffff32', margin: '7%' }}>
+          DID social recovery
+        </h2>
         <DidSocialRecovery />
       </Layout>
     </>
