+++ pages/[username]/xwallet/did/styles.module.scss
@@ -1,7 +1,7 @@
 @import url("https://fonts.googleapis.com/css?family=Noto+Sans+Mono");
 
 .headlineWrapper {
-  width: 100%;
+  width: 45%;
   text-align: center;
   margin-top: 7%;
   margin-bottom: 7%;
